+++ CHANGES.rst
@@ -26,6 +26,8 @@ New:
 
 - Add filter and search tiles.
 
+- Add a ``sort_key_function`` key to the IQueryModifier dict to allow for a different sort key function when sorting the values.
+
 - Add a ``index_modifier`` key to the IQueryModifier indexes dict to allow transforming of index search values.
   For ``KeywordIndex`` indices the index_modifier is automatically set to encode the value to utf-8.